training/requirements.txt
torch==2.0.1
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.2.2
tensorboard==2.17.0
plotly==5.15.0
pyyaml
